[fix] : 빌드 에러 수정
